fix nondottedpath settings
